restructure integration tests and create utils to reduce duped code
