--- integration/cmd/unresponsive_server/main.go
@@ -1,18 +1,0 @@
-package main
-
-import (
-	"bufio"
-	"os"
-)
-
-func main() {
-	reader := bufio.NewReader(os.Stdin)
-	go func() {
-		for {
-			buf := make([]byte, 1024)
-			reader.Read(buf)
-		}
-	}()
-
-	select {}
-}
