+++ internal/integration/config/otel_config.yaml
