+++ internal/integration/config/otel_grpc_lspwatch.yaml
