+++ internal/integration/config/otel_http_lspwatch.yaml
