+++ internal/integration/testdata/client_messages.json
