+++ internal/testutil/utills.go
@@ -1,11 +1,12 @@
 package testutil
 
 import (
+	"fmt"
+	"os"
+	"os/exec"
 	"testing"
 	"time"
 )
-
-// TODO: Tests.
 
 // AssertExitsAfter returns true if the function terminates after the timeout.
 // WARNING: The goroutine running fn will not be terminated when this function returns.
@@ -48,4 +49,27 @@ func AssertDoesNotExitBefore(t *testing.T, desc string, fn func(), timeout time.
 	case <-deadline:
 		return true
 	}
+}
+
+func RunIntegrationTest(t *testing.T, args []string, fn func(cmd *exec.Cmd)) {
+	lspwatchBinary := os.Getenv("LSPWATCH_BIN")
+	if lspwatchBinary == "" {
+		t.Fatalf("LSPWATCH_BIN is not set")
+	}
+
+	coverageDir := os.Getenv("COVERAGE_DIR")
+	if coverageDir == "" {
+		t.Fatalf("COVERAGE_DIR is not set")
+	}
+
+	cmd := exec.Command(
+		lspwatchBinary,
+		args...,
+	)
+
+	coverPath := fmt.Sprintf("GOCOVERDIR=%s", coverageDir)
+	t.Logf("Sending coverage data to: %s", coverageDir)
+	cmd.Env = append(os.Environ(), coverPath)
+
+	fn(cmd)
 }
